\name{checkfac}
\alias{checkfac}
\title{Check factor}
\description{
  Convert input \code{\link{vector}} into a \code{\link{factor}} suitable for analysis
  in ADE4. This function is called by \code{link[made4:bga]{bga}} and other made4
  function, it is not normally called by the user 
}
\usage{
checkfac(classvec)
}
\arguments{
  \item{classvec}{A \code{\link{vector}} or \code{\link{factor}}}
}
\details{
}
\value{
  \item{factor }{\code{\link{factor}} suitable for input into \code{bga}
  }
  }
\author{Aedin Culhane}
\note{  }
\seealso{ }
\examples{
}
\keyword{internal}% at least one, from doc/KEYWORDS

